requirements.txt
Django==4.2.7
django-ninja==1.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
dj-database-url==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
